Order Boards in BoardGroups deterministically
Previously, boards were sorted by the order in which they were
discovered. The discovery order should be considered
nondeterministic, since it depends on things like USB addresses
for hardware backends and is currently undefined for non-hardware
backends.

In an effort to make this more useful to a user, boards will now
be ordered deterministically by sorting the collection by board
serial number.

This order is only visible when iterating through the BoardGroup,
since #25 removed the ability to index boards by integer indices.

Closes #37.