Fix typing of MockUSBPowerBoardDevice.ctrl_transfer
The typechecker was complaining that `Optional[Union[int, bytes]]` was being
implicitly converted to `Union[int, bytes]`. After inspecting this code it
seems more sensible to dynamically check the argument type at the highest
level we can, and pass simply an `int` or `bytes` down to the helper
functions (`read_data` and `write_data`) as appropriate.