TEST PR DO NOT MERGE
This PR contains the same content as #254; however it is based
off an old revision of master from before #233 was merged.
Right now I expect this to pass CI because this new board agrees
with the rest of the codebase in this branch. However, merging
this with master pulls in #233 and hence causes CI to fail. We
need to catch this before the PR gets merged.