BoardGroup now takes two type parameters
This resolves the issue where mypy was asking for multiple type
parameters. It introduces a new helper function
`BoardGroup.get_board_group`, to help us create `BoardGroup` objects
without repeating the `Backend` type twice. This is needed because we
need to use the value for the `Backend` class, as well as the type.