Fix potential bug in MCv4B thread lock
The previous implementation acquired the lock twice, separately within
the firmware version command. This could potentially allow a situation
where a command had been issued to the board and another thread grabbed
the lock before we had read the data in the serial buffer. This could
obviously cause issues.

One of the other solutions that I tried was:

```python
    with self._lock:
        self.send_command(...)
        self._serial.read_line(...)
```

This results in an interesting situation which I shall leave as an
exercise to the reader to work out what goes wrong.

(Hint: The unit tests never complete. Not even an error :P)