Add flake8-print
Makefile is modified to run the `extracode.flake8` config against the
extracode. This is so that we can allow `print` statements in them,
without having to add `# noqa: T002` on almost every line.

A couple of lines are ignored in the console code, this is due to the
actual permitted use of `print` here.