backends/ruggeduino: Set the pin mode when writing digital outputs
This patch does set the pin mode on every output, which is identical in behaviour to how the SB Arduino backend also works.

As the _update_digital_pins function does not know the previous state of the pin, we cannot know when we need to change it,
therefore we set it every single time.

Additionally, when the pin mode is set to digital output, the pin will also receive a write call, as the state in the j5 code
will already contain a state for the pin (LOW by default). This ensures that the state of the pin, matches the state in j5.
It is intended behaviour.

Fixes #758

Co-Authored-By: Karina Kwiatek <kkwiatek@studentrobotics.org>