Distinguish between blocking and non-blocking buzzes
Some devices with a piezo buzzer will buzz without blocking the code, which is quite useful.

However, sometimes with these devices it is useful to be able to buzz and then have the code wait for the buzz to finish. For example, when playing a tune.

Each `Piezo` component instance should have a default blocking state, which can then be overriden by users when `piezo.buzz` is called.

In the case that a backend does not support blocking buzzes, a sleep function should be added.

In the case that a backend does not support non-blocking buzzes, it should raise an exception or spawn a background thread for the buzz.

Fixes #160